feat: Transfer functionality for external transactions

+++ src/models/Transactions.model.ts
@@ -530,64 +530,22 @@ export interface UpdateTransactionByIdParams {
 }
 
 export const updateTransactionById = async (
-  {
-    id,
-    userId,
-    amount,
-    refAmount,
-    note,
-    time,
-    transactionType,
-    paymentType,
-    accountId,
-    categoryId,
-    currencyId,
-    currencyCode,
-    refCurrencyCode,
-    isTransfer,
-    transferId,
-  }: UpdateTransactionByIdParams,
+  { id, userId, ...payload }: UpdateTransactionByIdParams,
   { transaction }: { transaction?: Transaction } = {},
 ) => {
   const where = { id, userId };
-  await Transactions.update(
-    {
-      amount,
-      refAmount,
-      note,
-      time,
-      transactionType,
-      paymentType,
-      accountId,
-      categoryId,
-      currencyCode,
-      refCurrencyCode,
-      isTransfer,
-      transferId,
-      currencyId,
-    },
-    {
-      where,
-      transaction,
-      individualHooks: true,
-    },
-  );
+
+  await Transactions.update(removeUndefinedKeys(payload), {
+    where,
+    transaction,
+    individualHooks: true,
+  });
 
   return getTransactionById({ id, userId }, { transaction });
 };
 
 export const updateTransactions = (
-  {
-    amount,
-    note,
-    time,
-    transactionType,
-    paymentType,
-    accountId,
-    categoryId,
-    currencyId,
-    refCurrencyCode,
-  }: {
+  payload: {
     amount?: number;
     note?: string;
     time?: Date;
@@ -603,17 +561,7 @@ export const updateTransactions = (
   { transaction }: { transaction?: Transaction } = {},
 ) => {
   return Transactions.update(
-    {
-      amount,
-      note,
-      time,
-      transactionType,
-      paymentType,
-      accountId,
-      categoryId,
-      currencyId,
-      refCurrencyCode,
-    },
+    removeUndefinedKeys(payload),
     {
       where,
       transaction,
@@ -622,10 +570,16 @@ export const updateTransactions = (
   );
 };
 
-export const deleteTransactionById = (
+export const deleteTransactionById = async (
   { id, userId }: { id: number; userId: number },
   { transaction }: { transaction?: Transaction } = {},
 ) => {
+  const tx = await getTransactionById({ id, userId }, { transaction });
+
+  if (tx.accountType !== ACCOUNT_TYPES.system) {
+    throw new ValidationError({ message: "It's not possible to manually delete external transactions" });
+  }
+
   return Transactions.destroy({
     where: { id, userId },
     transaction,
